requirements.txt
streamlit
openai-whisper
torch
langchain
langchain-core
langchain-groq
